+++ app/Url.php
@@ -2,81 +2,33 @@
 
 class Url
 {
-    protected $host;
+    protected $public_url;
 
-    protected $root;
+    protected $base_url;
 
-    protected $scriptName;
-
-    public function __construct()
+    public function __construct($base_url = null)
     {
-        $this->getHost();
-    }
-
-    protected function getHost()
-    {
-        if (!$this->isCommandLineInterface()) {
-            $protocol = isset($_SERVER["HTTPS"]) ? 'https' : 'http';
-            $this->host = $protocol . '://' . $_SERVER['HTTP_HOST'];
-        }
-        $this->getRoot();
-    }
-
-    protected function getRoot()
-    {
-        if ($this->isCommandLineInterface()) {
-            $this->host .= '/';
+        if (!empty($base_url)) {
+            $this->base_url = $base_url;
         } else {
-            $this->removeIndex();
-            $this->host .= $this->scriptName . "/";
+            $request = \App\Component\Request::getInstance();
+            $this->base_url = $request->getSchemeAndHttpHost() . $request->getBaseUrl();
         }
-    }
-
-    protected function removeIndex()
-    {
-        $this->scriptName = preg_replace('/\/index.php/', '', $_SERVER['SCRIPT_NAME']);
+        $this->public_url = $this->base_url . '/public/';
     }
 
     public function get($path = null)
     {
-        if (!$this->isCommandLineInterface()) {
-            if (preg_match('/index.php/', $_SERVER['REQUEST_URI'])) {
-                $this->host .= 'index.php/';
-            }
-        }
-        $path = trim(trim($path), '/');
-        $pathArray = explode('/', $path);
-
-        if (!$this->isCommandLineInterface()) {
-            if (end($pathArray) == "index") {
-                array_pop($pathArray);
-            }
-        }
-
-        $path = implode("/", $pathArray);
-
-        if ($this->isCommandLineInterface()) {
-            $path .= '.html';
-        }
-        return $this->host .= $path;
+        return $this->base_url . $path;
     }
 
     /**
-     * @example "asset('css/style.css')" http://localhost/css/style.css
+     * @example "asset('css/style.css')" http://localhost/public/css/style.css
      * @param null $content /public
      * @return string
      */
     public function asset($content = null)
     {
-        if ($this->isCommandLineInterface()) {
-            return '/' . trim(trim($content), '/');
-        } else {
-            return $this->host . 'public/' . trim(trim($content), '/');
-        }
-    }
-
-    private function isCommandLineInterface()
-    {
-        return (php_sapi_name() === 'cli');
+        return $this->public_url . trim(trim($content), '/');
     }
 }